Rename and deprecate configuration symbols
In order to make a distinction between global configurations and local
configurations, we use names at the top level which correspond to these
categories; hence global-configuration and local-configuration.

Additionally, to prevent potential name collisions in the future, numerous field
getters are renamed to use the full record name as a prefix.

With the exception of this-* symbols which is a special case that seems too
complicated to be worth maintaining compatibility for, deprecation warnings are
used for all replaced symbols so that the shift to the new names may be gradual.