Add support for remote name prefixes
By default when using local-configuration, remote names will be prefixed with
the string "auto-".  In addition to allowing other user-configured remotes to
coexist, this enables a pseudo-workaround for an issue with running enableremote
on git-annex borg remotes previously initialized in the same repo and then
removed via `git remote remove' (to-be-documented).